+++ src/Ampersand/Input/ADL1/Parser.hs
@@ -14,7 +14,7 @@ import           Ampersand.Core.ParseTree
 import           Ampersand.Input.ADL1.ParsingLib
 import           Data.List
 import qualified Data.Set as Set
-import qualified Data.List.NonEmpty as NEL (NonEmpty(..))
+import qualified Data.List.NonEmpty as NEL
 import           Data.Maybe
 
 --- Populations ::= Population+
@@ -42,7 +42,7 @@ pContext  = rebuild <$> posOf (pKey "CONTEXT")
             , ctx_rs     = [p | CRul p<-ces]       -- All user defined rules in this context, but outside patterns
             , ctx_ds     = [p | CRel p<-ces]       -- The relations defined in this context, outside the scope of patterns
             , ctx_cs     = [c ("CONTEXT "++nm) | CCon c<-ces]    -- The concept definitions defined in this context, outside the scope of patterns
-            , ctx_gs     = [g | CGen g<-ces] ++ [y | CCfy y<-ces] -- The gen definitions defined in this context, outside the scope of patterns
+            , ctx_gs     = concat [ys | CCfy ys<-ces]       -- The Classify definitions defined in this context, outside the scope of patterns
             , ctx_ks     = [k | CIndx k<-ces]      -- The identity definitions defined in this context, outside the scope of patterns
             , ctx_rrules = [x | Cm x <-ces]        -- The MAINTAINS statements in the context
             , ctx_rrels  = [x | Cl x <-ces]        -- The EDITS statements in the context
@@ -56,7 +56,7 @@ pContext  = rebuild <$> posOf (pKey "CONTEXT")
        , [s | CIncl s<-ces] -- the INCLUDE filenames
        )
 
-    --- ContextElement ::= Meta | PatternDef | ProcessDef | RuleDef | Classify | RelationDef | ConceptDef | GenDef | Index | ViewDef | Interface | Sqlplug | Phpplug | Purpose | Population | PrintThemes | IncludeStatement
+    --- ContextElement ::= Meta | PatternDef | ProcessDef | RuleDef | Classify | RelationDef | ConceptDef | Index | ViewDef | Interface | Sqlplug | Phpplug | Purpose | Population | PrintThemes | IncludeStatement
     pContextElement :: AmpParser ContextElement
     pContextElement = CMeta    <$> pMeta         <|>
                       CPat     <$> pPatternDef   <|>
@@ -68,7 +68,6 @@ pContext  = rebuild <$> posOf (pKey "CONTEXT")
                       Cm       <$> pRoleRule     <|>
                       Cm       <$> pServiceRule  <|>
                       Cl       <$> pRoleRelation <|>
-                      CGen     <$> pGenDef       <|>
                       CIndx    <$> pIndex        <|>
                       CView    <$> pViewDef      <|>
                       Cifc     <$> pInterface    <|>
@@ -79,13 +78,12 @@ pContext  = rebuild <$> posOf (pKey "CONTEXT")
 data ContextElement = CMeta Meta
                     | CPat P_Pattern
                     | CRul (P_Rule TermPrim)
-                    | CCfy P_Gen
+                    | CCfy [PClassify]
                     | CRel P_Relation
                     | CCon (String -> ConceptDef)
                     | CRep Representation
                     | Cm P_RoleRule
                     | Cl P_RoleRelation
-                    | CGen P_Gen
                     | CIndx P_IdentDef
                     | CView P_ViewDef
                     | Cifc P_Interface
@@ -138,7 +136,7 @@ pPatternDef' (beginKeyword,endKeyword)
      = P_Pat { pos = pos'
              , pt_nm  = nm
              , pt_rls = [r | Pr r<-pes]
-             , pt_gns = [y | Py y<-pes] ++ [g | Pg g<-pes]
+             , pt_gns = concat [ys | Py ys<-pes]
              , pt_dcs = [d | Pd d<-pes]
              , pt_RRuls = [rr | Pm rr<-pes]
              , pt_RRels = [rr | Pl rr<-pes]
@@ -152,7 +150,7 @@ pPatternDef' (beginKeyword,endKeyword)
              }
 
 -- PatElem used by PATTERN and PROCESS
---- PatElem ::= RuleDef | Classify | RelationDef | ConceptDef | GenDef | Index | ViewDef | Purpose | Population
+--- PatElem ::= RuleDef | Classify | RelationDef | ConceptDef | Index | ViewDef | Purpose | Population
 pPatElem :: AmpParser PatElem
 pPatElem = Pr <$> pRuleDef          <|>
            Py <$> pClassify         <|>
@@ -163,7 +161,6 @@ pPatElem = Pr <$> pRuleDef          <|>
            Pm <$> pServiceRule      <|>
            Pl <$> pRoleRelation     <|>
            Pc <$> pConceptDef       <|>
-           Pg <$> pGenDef           <|>
            Prep <$> pRepresentation <|>
            Pk <$> pIndex            <|>
            Pv <$> pViewDef          <|>
@@ -171,31 +168,43 @@ pPatElem = Pr <$> pRuleDef          <|>
            Pp <$> pPopulation
 
 data PatElem = Pr (P_Rule TermPrim)
-             | Py P_Gen
+             | Py [PClassify]
              | Pd P_Relation
              | Pm P_RoleRule
              | Pl P_RoleRelation
              | Pc (String -> ConceptDef)
-             | Pg P_Gen
              | Prep Representation
              | Pk P_IdentDef
              | Pv P_ViewDef
              | Pe PPurpose
              | Pp P_Population
 
---- Classify ::= 'CLASSIFY' ConceptRef 'IS' Cterm
-pClassify :: AmpParser P_Gen   -- Example: CLASSIFY A IS B /\ C /\ D
-pClassify = try (P_Cy <$> currPos
-                      <* pKey "CLASSIFY"
-                      <*> pConceptRef
-                      <*  pKey "IS")
-                 <*> pCterm
+--- Classify ::= 'CLASSIFY' ConceptRef ('IS' Cterm | 'ISA' ConceptRef)
+pClassify :: AmpParser [PClassify]   -- Example: CLASSIFY A IS B /\ C /\ D
+pClassify = fun <$> currPos
+                <*  pKey "CLASSIFY"
+                <*> pConceptRef `sepBy1` pComma
+                <*> (     (is  <$ pKey "IS"  <*> pCterm)
+                      <|> (isa <$ pKey "ISA" <*> pConceptRef)
+                    )
                where
+                 fun :: Origin -> [P_Concept] -> (Bool, [P_Concept]) -> [PClassify]
+                 fun p lhs (isISA ,rhs) = map f lhs
+                   where 
+                     f s = PClassify 
+                             { pos      = p
+                             , specific = s
+                             , generics = if isISA then s NEL.:| rhs else NEL.fromList rhs
+                             }
                  --- Cterm ::= Cterm1 ('/\' Cterm1)*
                  --- Cterm1 ::= ConceptRef | ('('? Cterm ')'?)
                  pCterm  = concat <$> pCterm1 `sepBy1` pOperator "/\\"
                  pCterm1 = pure   <$> pConceptRef <|>
                                       pParens pCterm  -- brackets are allowed for educational reasons.
+                 is :: [P_Concept] -> (Bool, [P_Concept])
+                 is gens = (False, gens)
+                 isa :: P_Concept -> (Bool, [P_Concept])
+                 isa gen = (True, [gen])
 
 --- RuleDef ::= 'RULE' Label? Rule Meaning* Message* Violation?
 pRuleDef :: AmpParser (P_Rule TermPrim)
@@ -332,11 +341,6 @@ pAdlTType
   where
    k tt str = f <$> pKey str where f _ = tt
 
---- GenDef ::= ('CLASSIFY' | 'SPEC') ConceptRef 'ISA' ConceptRef
-pGenDef :: AmpParser P_Gen
-pGenDef = try (PGen <$> currPos <* key <*> pConceptRef <* pKey "ISA") <*> pConceptRef --
-          where key = pKey "CLASSIFY" <|> pKey "SPEC"
-
 -- | A identity definition looks like:   IDENT onNameAdress : Person(name, address),
 -- which means that name<>name~ /\ address<>addres~ |- I[Person].
 -- The label 'onNameAddress' is used to refer to this identity.
@@ -408,16 +412,18 @@ pViewDefLegacy = P_Vd <$> currPos
 --- Interface ::= 'INTERFACE' ADLid Params? Roles? ':' Term (ADLid | Conid)? SubInterface?
 pInterface :: AmpParser P_Interface
 pInterface = lbl <$> currPos                                       
-                 <*> (pInterfaceKey *> pADLid)
+                 <*> pInterfaceIsAPI
+                 <*> pADLid
                  <*> optList pParams
                  <*> optList pRoles 
                  <*> (pColon *> pTerm)          -- the expression of the interface object
                  <*> pMaybe pCruds              -- The Crud-string (will later be tested, that it can contain only characters crud (upper/lower case)
                  <*> pMaybe (pChevrons pConid)  -- The view that should be used for this object
                  <*> pSubInterface
-    where lbl :: Origin -> String ->  [P_NamedRel] -> [Role] -> Term TermPrim -> Maybe P_Cruds -> Maybe String -> P_SubInterface -> P_Interface
-          lbl p nm _params roles ctx mCrud mView sub
-             = P_Ifc { ifc_Name   = nm
+    where lbl :: Origin -> Bool -> String ->  [P_NamedRel] -> [Role] -> Term TermPrim -> Maybe P_Cruds -> Maybe String -> P_SubInterface -> P_Interface
+          lbl p isAPI nm _params roles ctx mCrud mView sub
+             = P_Ifc { ifc_IsAPI  = isAPI
+                     , ifc_Name   = nm
                      , ifc_Roles  = roles
                      , ifc_Obj    = P_BxExpr { obj_nm   = nm
                                           , pos      = p
@@ -521,6 +527,9 @@ pPurpose = rebuild <$> currPos
 
 pInterfaceKey :: AmpParser String
 pInterfaceKey = pKey "INTERFACE" <|> pKey "API" -- On special request of Rieks, the keyword "API" is allowed everywhere where the keyword "INTERFACE" is used. https://github.com/AmpersandTarski/Ampersand/issues/789
+
+pInterfaceIsAPI :: AmpParser Bool
+pInterfaceIsAPI = ("API" ==) <$> pInterfaceKey
 
 --- Population ::= 'POPULATION' (NamedRel 'CONTAINS' Content | ConceptName 'CONTAINS' '[' ValueList ']')
 -- | Parses a population