+++ ReleaseNotes.md
@@ -1,5 +1,10 @@
 # Release notes of Ampersand
 
+## v3.12.0 (21 december 2018)
+
+* [Issue #855](https://github.com/AmpersandTarski/Ampersand/issues/855) Minor enhancement in CLASSIFY statement
+* Add semantics to difference in API or INTERFACE keywords.
+
 ## v3.11.5 (23 november 2018)
 
 * [Issue #628](https://github.com/AmpersandTarski/Ampersand/issues/628) Fixed a performance issue for specific queries.